add confirm to delete single product method

+++ src/components/DeleteButton.tsx
@@ -19,16 +19,18 @@ const DeleteButton = ({ id }: { id: string }) => {
   }
 
   const handleDelete = async () => {
-    const res = await fetch(`http://localhost:3000/api/products/${id}`, {
-      method: "DELETE",
-    });
+    if (confirm("Are you sure you want to delete this product ?")) {
+      const res = await fetch(`http://localhost:3000/api/products/${id}`, {
+        method: "DELETE",
+      });
 
-    if (res.status === 200) {
-      router.push("/menu");
-      toast.success("The product has been deleted!");
-    } else {
-      const data = await res.json();
-      toast.error(data.message);
+      if (res.status === 200) {
+        router.push("/menu");
+        toast.success("The product has been deleted!");
+      } else {
+        const data = await res.json();
+        toast.error(data.message);
+      }
     }
   };
